# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KnowFlow AI",
    "description": "Enterprise Knowledge & SOP Agent backend API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/stats": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Admin Stats",
        "description": "Returns aggregated system metrics across documents, queries, and feedback.",
        "operationId": "get_admin_stats_api_admin_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/audit-logs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Audit Logs",
        "description": "Returns paginated audit trails providing auditability foundations relevant to regulated environments.",
        "operationId": "get_audit_logs_api_admin_audit_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLogItem"
                  },
                  "title": "Response Get Audit Logs Api Admin Audit Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/stats": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Admin Stats",
        "description": "Returns aggregated system metrics across documents, queries, and feedback.",
        "operationId": "get_admin_stats_api_v1_admin_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/audit-logs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Audit Logs",
        "description": "Returns paginated audit trails providing auditability foundations relevant to regulated environments.",
        "operationId": "get_audit_logs_api_v1_admin_audit_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLogItem"
                  },
                  "title": "Response Get Audit Logs Api V1 Admin Audit Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminStatsResponse": {
        "properties": {
          "total_documents": {
            "type": "integer",
            "title": "Total Documents"
          },
          "total_chunks": {
            "type": "integer",
            "title": "Total Chunks"
          },
          "total_conversations": {
            "type": "integer",
            "title": "Total Conversations"
          },
          "total_messages": {
            "type": "integer",
            "title": "Total Messages"
          },
          "total_audit_events": {
            "type": "integer",
            "title": "Total Audit Events"
          },
          "positive_feedback_count": {
            "type": "integer",
            "title": "Positive Feedback Count"
          },
          "negative_feedback_count": {
            "type": "integer",
            "title": "Negative Feedback Count"
          },
          "grounded_answering_policy": {
            "type": "string",
            "title": "Grounded Answering Policy",
            "default": "Active"
          },
          "hybrid_reranker_status": {
            "type": "string",
            "title": "Hybrid Reranker Status",
            "default": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "total_documents",
          "total_chunks",
          "total_conversations",
          "total_messages",
          "total_audit_events",
          "positive_feedback_count",
          "negative_feedback_count"
        ],
        "title": "AdminStatsResponse"
      },
      "AuditLogItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "metadata_json": {
            "type": "string",
            "title": "Metadata Json"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "action",
          "metadata_json",
          "created_at"
        ],
        "title": "AuditLogItem"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}